Initial commit
This is "confirmed working" in another project, and was ported to a
separate project so that it could be used in more than one place.

I use quotes for "confirmed working" because the *decoding* parts are
definitely all correct but the *encoding* bits have not been
well-tested. It shoudl be somewhat trivial to write up some test cases
and get better confidence on the correctness of this library.