Change name to avoid collision with existing package
A little upset I waited as long as I did to do this. Had this code
mostly ready to go back in January and I just got super busy until one
week after someone else claimed the name. Oh well, I just added `_ex` to
the end of the name and it communicates the specific language (elixir)
that the source code uses, so it's not the worst ever.